# Use a size_stats table