Don't know why I didn't use a table for this one; it's a massive performance
win.

Before when selecting a month:

	backend  1219ms  pages.List
	backend     3ms  browsers.List
	backend   554ms  sizeStat.ListSizes
	backend     6ms  locStat.List
	backend     2ms  topRefs.List
	backend    13ms  zhttp.Template

After:

	backend  1020ms  pages.List
	backend     5ms  browsers.List
	backend     2ms  sizeStat.ListSizes
	backend     5ms  locStat.List
	backend     5ms  topRefs.List
	backend    33ms  zhttp.Template